// src/Enums/Pages.ts
export enum Pages {
    SignIn = "/SignIn",
    ClientPage = "/MainPage/Clients",
    Devices = "/MainPage/Devices",
    MainPage = "/MainPage",
    SignUp = "/SignUp",
    NotFound = "/404",
    Account = 'account',
    Profile = "/Profile",
    Sensors = '/MainPage/Sensors',
    ClientChart = "/MainPage/Chart",
    Measurement = "/MainPage/Measurement"
}

export enum Client {

}